requirements.txt
email-validator==1.1.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
Pillow==7.2.0
PyJWT==1.7.1
python-dotenv==0.14.0
python-slugify==4.0.1
